refactor: creating an LLM module in libmemex and moving embedding module inside
