+++ lib/libmemex/src/llm/embedding.rs
