+++ lib/libmemex/src/llm/mod.rs
@@ -1,2 +1,3 @@
+pub mod embedding;
 pub mod openai;
 pub mod prompter;
